GetExpBackoff from pod to util
